webapp: one Map page — trips as first-class objects, chart<->map crosshair
Live and History were a false split: both were "a map of a day", one pinned
to now. Merged into a single Map page where the SHOW chips pick the view:

  LIVE   (today only) follow the device — the current/latest trip plus the
         blue dot, auto-refreshing with INCREMENTAL fetches (from_ts past the
         last-seen fix; a 1 Hz day is far too big to re-download every 5 s).
  DAY    the whole day.
  T1..Tn detected trips, Garmin-style: a trip is a run of GPS fixes with no
         gap >5 min and no stationary spell >2 min (Doppler <0.8 m/s and step
         <10 m); runs under 60 s / 80 m are noise. Segmentation is
         client-side (trips.js), so rules can evolve without touching data.

The timeline keeps its drag handles (custom window -> "range" selection) and
gains clickable trip bands; isolating two 5-minute rides 10 minutes apart is
now one click each. Chart hover snaps to nearest-x, shows a dot, and rings
the same point on the map; table rows do the same on hover.

1 Hz volume work that this forced: density strip is a canvas (was one div
per fix), charts decimate above 1500 points (each kept point references its
original fix so hover stays exact), day loads raise the API limit, and
/api/positions accepts float from_ts/to_ts (received_ts is REAL since v3).

Verified: all modules parse; trip segmentation checked against 5 scenarios
(two rides 10 min apart -> 2 trips; 60 s traffic light -> no split; 3 min
parked mid-sampling -> split; parked day / 40 s errand -> none); static +
API routes serve; smoke.sh PASS. Interactive behaviour needs a browser look.

+++ server/app.py
@@ -52,7 +52,7 @@ async def devices():
 
 @app.get("/api/positions")
 async def positions(device: str, since: int = 0,
-                    from_ts: int = 0, to_ts: int = 0, limit: int = 5000):
+                    from_ts: float = 0, to_ts: float = 0, limit: int = 5000):
     """Position history for one device, chronological.
 
     Live uses `since` (minutes back from now). Replay uses absolute epoch
